new big change to support binaries

+++ gterm/shell.py
@@ -14,7 +14,9 @@ import bashlex
 import gterm.colorshell as gshc
 import importlib
 import json
+import gterm.gterm_filesystem as gfs
 import requests
+import subprocess
 
 # ripped from gkux terminal
 cprint = lambda text: gshc.print_with_name(text, "gsh")
@@ -23,23 +25,29 @@ def start_shell():
     cprint("welcome to gterminal!")
     while True:
         try:
-            raw_command = input(":-$ ")
+            raw_command = input(":-$ ").strip()
             if (raw_command.startswith("#")): continue
+            if (len(raw_command) == 0): continue
             argv = list(bashlex.split(raw_command))
-            c = importlib.import_module("system.scripts." + argv[0])
-            c.onExecute(argv)
+            if (gfs.exists("/scripts/" + argv[0] + ".py")):
+                c = importlib.import_module("system.scripts." + argv[0])
+                c.onExecute(argv)
+            elif gfs.exists("/bin/" + argv[0]):
+                subprocess.run([gfs.getsysroot("/bin/" + argv[0])] + argv[1:], shell=True)
+            else:
+                missing = requests.get("https://raw.githubusercontent.com/thekaigonzalez/gterminal/master/software/MISSING_SOFTWARE.json")
+                js = (missing.json())
+                if (js.get(argv[0]) != None):
+                    cprint("that command can be found in the package '{}'.".format(js.get(argv[0])))
+                    cprint("to install it: `pkg get " + js[argv[0]] + "'")
+                else:
+                    cprint("that command, {}, does not exist. sorry".format(argv[0]))
         except KeyboardInterrupt: # ^C
-            cprint("press ^D to exit.")
+            print()
+            pass
         except EOFError: # ^D
             exit(0)
-        except ModuleNotFoundError:
-            missing = requests.get("https://raw.githubusercontent.com/thekaigonzalez/gterminal/master/software/MISSING_SOFTWARE.json")
-            js = (missing.json())
-            if (js.get(argv[0]) != None):
-                cprint("that command can be found in the package '{}'.".format(js.get(argv[0])))
-                cprint("to install it: `pkg get " + js[argv[0]] + "'")
-            else:
-                cprint("that command, {}, does not exist. sorry".format(argv[0]))
+        
         except Exception as e:
             cprint("there was an error running your program.")
             cprint("error: " + str(e))
